+++ channel.go
@@ -117,7 +117,6 @@ type Channel struct {
 	createdStack      string
 	commonStatsTags   map[string]string
 	connectionOptions ConnectionOptions
-	handlers          *handlerMap
 	peers             *PeerList
 
 	// mutable contains all the members of Channel which are mutable.
@@ -190,7 +189,6 @@ func NewChannel(serviceName string, opts *ChannelOptions) (*Channel, error) {
 		},
 
 		connectionOptions: opts.DefaultConnectionOptions,
-		handlers:          &handlerMap{},
 	}
 	ch.peers = newRootPeerList(ch).newChild()
 
@@ -297,9 +295,16 @@ type Registrar interface {
 	Peers() *PeerList
 }
 
-// Register registers a handler for a service+method pair
+// Register registers a handler for a method.
+//
+// The handler is registered with the service name used when the Channel was
+// created. To register a handler with a different service name, obtain a
+// SubChannel for that service with GetSubChannel, and Register a handler
+// under that. You may also use SetHandler on a SubChannel to set up a
+// catch-all Handler for that service. See the docs for SetHandler for more
+// information.
 func (ch *Channel) Register(h Handler, methodName string) {
-	ch.handlers.register(h, ch.PeerInfo().ServiceName, methodName)
+	ch.GetSubChannel(ch.PeerInfo().ServiceName).Register(h, methodName)
 }
 
 // PeerInfo returns the current peer info for the channel
@@ -550,7 +555,7 @@ func (ch *Channel) removeClosedConn(c *Connection) {
 func (ch *Channel) connectionCloseStateChange(c *Connection) {
 	ch.removeClosedConn(c)
 	if peer, ok := ch.rootPeers().Get(c.remotePeerInfo.HostPort); ok {
-		peer.connectionStateChanged(c)
+		peer.connectionCloseStateChange(c)
 		ch.updatePeer(peer)
 	}