+++ peer_bench_test.go
@@ -21,54 +21,40 @@
 package tchannel_test
 
 import (
-	"runtime"
 	"testing"
 	"time"
 
 	. "github.com/uber/tchannel-go"
 
+	"github.com/stretchr/testify/require"
 	"github.com/uber/tchannel-go/testutils"
 )
 
-func waitForChannelClose(t *testing.T, ch *Channel) bool {
-	started := time.Now()
+func benchmarkGetConnection(b *testing.B, numIncoming, numOutgoing int) {
+	ctx, cancel := NewContext(10 * time.Second)
+	defer cancel()
 
-	var state ChannelState
-	for i := 0; i < 50; i++ {
-		if state = ch.State(); state == ChannelClosed {
-			return true
-		}
+	s1 := testutils.NewServer(b, nil)
+	s2 := testutils.NewServer(b, nil)
+	defer s1.Close()
+	defer s2.Close()
 
-		runtime.Gosched()
-		if i < 5 {
-			continue
-		}
-
-		sleepFor := time.Duration(i) * 100 * time.Microsecond
-		time.Sleep(testutils.Timeout(sleepFor))
+	for i := 0; i < numOutgoing; i++ {
+		_, err := s1.Connect(ctx, s2.PeerInfo().HostPort)
+		require.NoError(b, err, "Connect from s1 -> s2 failed")
+	}
+	for i := 0; i < numIncoming; i++ {
+		_, err := s2.Connect(ctx, s1.PeerInfo().HostPort)
+		require.NoError(b, err, "Connect from s2 -> s1 failed")
 	}
 
-	// Channel is not closing, fail the test.
-	sinceStart := time.Since(started)
-	t.Errorf("Channel did not close after %v, last state: %v", sinceStart, state)
-	return false
-}
-
-// WithVerifiedServer runs the given test function with a server channel that is verified
-// at the end to make sure there are no leaks (e.g. no exchanges leaked).
-func WithVerifiedServer(t *testing.T, opts *testutils.ChannelOpts, f func(serverCh *Channel, hostPort string)) {
-	var ch *Channel
-	testutils.WithServer(t, opts, func(serverCh *Channel, hostPort string) {
-		f(serverCh, hostPort)
-		ch = serverCh
-	})
-
-	if !waitForChannelClose(t, ch) {
-		return
-	}
-
-	// Check the message exchanges and make sure they are all empty.
-	if exchangesLeft := CheckEmptyExchanges(ch); exchangesLeft != "" {
-		t.Errorf("Found uncleared message exchanges:\n%v", exchangesLeft)
+	peer := s1.Peers().GetOrAdd(s2.PeerInfo().HostPort)
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		peer.GetConnection(ctx)
 	}
 }
+
+func BenchmarkGetConnection0In1Out(b *testing.B) { benchmarkGetConnection(b, 0, 1) }
+func BenchmarkGetConnection1In0Out(b *testing.B) { benchmarkGetConnection(b, 1, 0) }
+func BenchmarkGetConnection5In5Out(b *testing.B) { benchmarkGetConnection(b, 5, 5) }