+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+# 1.0.4
+
+* Improve handling of network failures during pending calls. Previously, calls
+  would timeout, but now they fail as soon as the network failure is detected.
+* Remove ephemeral peers with closed outbound connections.
+* #233: Ensure errors returned from Thrift handlers have a non-nil value.
+* #228: Add registered methods to the introspection output.
+* Add ability to set a global handler for a SubChannel.
+
 # 1.0.3
 
 * Improved performance when writing Thrift structs